+++ tests/integration/sugar/test_wallet.py
@@ -1,12 +1,12 @@
+import asyncio
 import time
+from threading import Thread
 
 from tests.integration.integration_test_case import IntegrationTestCase
-from tests.integration.it_utils import submit_transaction_async, test_async_and_sync
-from tests.integration.reusable_values import WALLET
+from tests.integration.it_utils import submit_transaction_async
 from xrpl.asyncio.clients import AsyncJsonRpcClient, AsyncWebsocketClient
 from xrpl.asyncio.wallet import generate_faucet_wallet
 from xrpl.clients import JsonRpcClient, WebsocketClient
-from xrpl.core.addresscodec import classic_address_to_xaddress
 from xrpl.models.requests import AccountInfo
 from xrpl.models.transactions import Payment
 from xrpl.wallet import generate_faucet_wallet as sync_generate_faucet_wallet
@@ -15,8 +15,12 @@ from xrpl.wallet.main import Wallet
 time_of_last_hooks_faucet_call = 0
 
 
-def sync_generate_faucet_wallet_and_fund_again(self, client, faucet_host=None):
-    wallet = sync_generate_faucet_wallet(client, faucet_host=faucet_host)
+def sync_generate_faucet_wallet_and_fund_again(
+    self, client, faucet_host=None, usage_context="integration_test"
+):
+    wallet = sync_generate_faucet_wallet(
+        client, faucet_host=faucet_host, usage_context=usage_context
+    )
     result = client.request(
         AccountInfo(
             account=wallet.classic_address,
@@ -25,7 +29,9 @@ def sync_generate_faucet_wallet_and_fund_again(self, client, faucet_host=None):
     balance = int(result.result["account_data"]["Balance"])
     self.assertTrue(balance > 0)
 
-    new_wallet = sync_generate_faucet_wallet(client, wallet, faucet_host=faucet_host)
+    new_wallet = sync_generate_faucet_wallet(
+        client, wallet, faucet_host=faucet_host, usage_context="integration_test"
+    )
     new_result = client.request(
         AccountInfo(
             account=new_wallet.classic_address,
@@ -35,8 +41,12 @@ def sync_generate_faucet_wallet_and_fund_again(self, client, faucet_host=None):
     self.assertTrue(new_balance > balance)
 
 
-async def generate_faucet_wallet_and_fund_again(self, client, faucet_host=None):
-    wallet = await generate_faucet_wallet(client, faucet_host=faucet_host)
+async def generate_faucet_wallet_and_fund_again(
+    self, client, faucet_host=None, usage_context="integration_test"
+):
+    wallet = await generate_faucet_wallet(
+        client, faucet_host=faucet_host, usage_context=usage_context
+    )
     result = await client.request(
         AccountInfo(
             account=wallet.classic_address,
@@ -45,7 +55,9 @@ async def generate_faucet_wallet_and_fund_again(self, client, faucet_host=None):
     balance = int(result.result["account_data"]["Balance"])
     self.assertTrue(balance > 0)
 
-    new_wallet = await generate_faucet_wallet(client, wallet, faucet_host=faucet_host)
+    new_wallet = await generate_faucet_wallet(
+        client, wallet, faucet_host=faucet_host, usage_context=usage_context
+    )
     new_result = await client.request(
         AccountInfo(
             account=new_wallet.classic_address,
@@ -56,19 +68,40 @@ async def generate_faucet_wallet_and_fund_again(self, client, faucet_host=None):
 
 
 class TestWallet(IntegrationTestCase):
-    @test_async_and_sync(
-        globals(),
-        ["xrpl.wallet.generate_faucet_wallet"],
-        num_retries=5,
-        use_testnet=True,
-    )
-    async def test_generate_faucet_wallet_rel_sub(self, client):
+    async def test_run_faucet_tests(self):
+        # run all the tests that start with `_test_` in parallel
+        def run_test(test_name):
+            with self.subTest(method=test_name):
+                method = getattr(self, test_name)  # get the test with the given name
+                if asyncio.iscoroutinefunction(method):  # is async
+                    asyncio.run(method())
+                else:  # is sync
+                    method()
+
+        # get all the test methods starting with `_parallel_test_`
+        # (the ones to run in parallel)
+        test_methods = [
+            method for method in dir(self) if method.startswith("_parallel_test_")
+        ]
+
+        # run all the tests in parallel
+        processes = []
+        for method in test_methods:
+            process = Thread(target=run_test, args=(method,))
+            process.start()
+            processes.append(process)
+        for process in processes:
+            process.join()
+
+    # ensure that the wallet creation has been validated and the account actually exists
+    async def _parallel_test_generate_faucet_wallet_rel_sub(self):
+        client = JsonRpcClient("https://s.devnet.rippletest.net:51234")
         destination = await generate_faucet_wallet(client)
         wallet = await generate_faucet_wallet(client)
+        # TODO: refactor so this actually waits for validation
         response = await submit_transaction_async(
             Payment(
                 account=wallet.classic_address,
-                sequence=wallet.sequence,
                 fee="10",
                 amount="1",
                 destination=destination.classic_address,
@@ -78,61 +111,83 @@ class TestWallet(IntegrationTestCase):
         )
         self.assertTrue(response.is_successful())
 
-    async def test_generate_faucet_wallet_testnet_async_websockets(self):
+    # Custom host tests
+
+    async def _parallel_test_generate_faucet_wallet_custom_host_async_websockets(self):
+        async with AsyncWebsocketClient(
+            "wss://s.devnet.rippletest.net:51233"
+        ) as client:
+            await generate_faucet_wallet_and_fund_again(
+                self,
+                client,
+                "faucet.devnet.rippletest.net",
+                usage_context="integration_test",
+            )
+
+    async def _parallel_test_generate_faucet_wallet_custom_host_async_json_rpc(self):
+        client = AsyncJsonRpcClient("https://s.devnet.rippletest.net:51234")
+        await generate_faucet_wallet_and_fund_again(
+            self,
+            client,
+            "faucet.devnet.rippletest.net",
+            usage_context="integration_test",
+        )
+
+    def _parallel_test_generate_faucet_wallet_custom_host_sync_websockets(self):
+        with WebsocketClient("wss://s.devnet.rippletest.net:51233") as client:
+            sync_generate_faucet_wallet_and_fund_again(
+                self, client, "faucet.devnet.rippletest.net"
+            )
+
+    def _parallel_test_generate_faucet_wallet_custom_host_sync_json_rpc(self):
+        client = JsonRpcClient("https://s.devnet.rippletest.net:51234")
+        sync_generate_faucet_wallet_and_fund_again(
+            self, client, "faucet.devnet.rippletest.net"
+        )
+
+    # Network tests
+
+    async def _parallel_test_generate_faucet_wallet_testnet_async_websockets(self):
         async with AsyncWebsocketClient(
             "wss://s.altnet.rippletest.net:51233"
         ) as client:
             await generate_faucet_wallet_and_fund_again(self, client)
 
-    async def test_generate_faucet_wallet_custom_host_async_websockets(self):
-        async with AsyncWebsocketClient(
-            "wss://s.devnet.rippletest.net:51233/"
-        ) as client:
-            await generate_faucet_wallet_and_fund_again(
-                self, client, "faucet.devnet.rippletest.net"
-            )
-
-    async def test_generate_faucet_wallet_custom_host_async_json_rpc(self):
-        client = AsyncJsonRpcClient("https://s.devnet.rippletest.net:51234/")
-        await generate_faucet_wallet_and_fund_again(
-            self, client, "faucet.devnet.rippletest.net"
-        )
-
-    def test_generate_faucet_wallet_custom_host_sync_websockets(self):
-        with WebsocketClient("wss://s.devnet.rippletest.net:51233/") as client:
-            sync_generate_faucet_wallet_and_fund_again(
-                self, client, "faucet.devnet.rippletest.net"
-            )
-
-    def test_generate_faucet_wallet_custom_host_sync_json_rpc(self):
-        client = JsonRpcClient("https://s.devnet.rippletest.net:51234/")
-        sync_generate_faucet_wallet_and_fund_again(
-            self, client, "faucet.devnet.rippletest.net"
-        )
-
-    async def test_generate_faucet_wallet_devnet_async_websockets(self):
+    async def _parallel_test_generate_faucet_wallet_devnet_async_websockets(self):
         async with AsyncWebsocketClient(
             "wss://s.devnet.rippletest.net:51233"
         ) as client:
             await generate_faucet_wallet_and_fund_again(self, client)
 
-    async def test_generate_faucet_wallet_amm_devnet_async_websockets(self):
+    async def _parallel_test_generate_faucet_wallet_amm_devnet_async_websockets(self):
         async with AsyncWebsocketClient(
             "wss://amm.devnet.rippletest.net:51233"
         ) as client:
             await generate_faucet_wallet_and_fund_again(self, client)
 
-    async def test_generate_faucet_wallet_hooks_v2_testnet_async_websockets(self):
+    async def _parallel_test_generate_faucet_wallet_sidechain_devnet_async_websockets(
+        self,
+    ):
         async with AsyncWebsocketClient(
-            "wss://hooks-testnet-v2.xrpl-labs.com"
+            "wss://sidechain-net1.devnet.rippletest.net:51233"
+        ) as client:
+            await generate_faucet_wallet_and_fund_again(self, client)
+
+    async def _parallel_test_generate_faucet_wallet_hooks_v3_testnet_async_websockets(
+        self,
+    ):
+        async with AsyncWebsocketClient(
+            "wss://hooks-testnet-v3.xrpl-labs.com"
         ) as client:
             global time_of_last_hooks_faucet_call
-            # Wait at least 10 seconds since last call to hooks v2 testnet faucet
+            # Wait at least 10 seconds since last call to hooks v3 testnet faucet
             time_since_last_hooks_call = time.time() - time_of_last_hooks_faucet_call
             if time_since_last_hooks_call < 10:
                 time.sleep(11 - time_since_last_hooks_call)
 
-            wallet = await generate_faucet_wallet(client)
+            wallet = await generate_faucet_wallet(
+                client, usage_context="integration_test"
+            )
             time_of_last_hooks_faucet_call = time.time()
 
             result = await client.request(
@@ -143,15 +198,20 @@ class TestWallet(IntegrationTestCase):
             balance = int(result.result["account_data"]["Balance"])
             self.assertTrue(balance > 0)
 
-    # Named different from test_generate_faucet_wallet_hooks_v2_testnet_async_websockets
-    # so the test runs far from each other since hooks v2 testnet faucet
+    # Named different from test_generate_faucet_wallet_hooks_v3_testnet_async_websockets
+    # so the test runs far from each other since hooks v3 testnet faucet
     # requires 10 seconds between calls
-    async def test_fund_given_wallet_hooks_v2_testnet_async_websockets(self):
+    async def test_fund_given_wallet_hooks_v3_testnet_async_websockets(self):
         async with AsyncWebsocketClient(
-            "wss://hooks-testnet-v2.xrpl-labs.com"
+            "wss://hooks-testnet-v3.xrpl-labs.com"
         ) as client:
             global time_of_last_hooks_faucet_call
-            wallet = Wallet("sEdSigMti9uJFCnrkwsB3LJRGkVZHVA", 0)
+            wallet = Wallet.create()
+            time_since_last_hooks_call = time.time() - time_of_last_hooks_faucet_call
+            if time_since_last_hooks_call < 10:
+                time.sleep(11 - time_since_last_hooks_call)
+            time_of_last_hooks_faucet_call = time.time()
+            await generate_faucet_wallet(client, wallet)
             result = await client.request(
                 AccountInfo(
                     account=wallet.classic_address,
@@ -159,13 +219,15 @@ class TestWallet(IntegrationTestCase):
             )
             balance = int(result.result["account_data"]["Balance"])
 
-            # Wait at least 10 seconds since last call to hooks v2 testnet faucet
+            # Wait at least 10 seconds since last call to hooks v3 testnet faucet
             time_since_last_hooks_call = time.time() - time_of_last_hooks_faucet_call
             if time_since_last_hooks_call < 10:
                 time.sleep(11 - time_since_last_hooks_call)
             time_of_last_hooks_faucet_call = time.time()
 
-            new_wallet = await generate_faucet_wallet(client, wallet)
+            new_wallet = await generate_faucet_wallet(
+                client, wallet, usage_context="integration_test"
+            )
             new_result = await client.request(
                 AccountInfo(
                     account=new_wallet.classic_address,
@@ -173,7 +235,3 @@ class TestWallet(IntegrationTestCase):
             )
             new_balance = int(new_result.result["account_data"]["Balance"])
             self.assertTrue(new_balance > balance)
-
-    def test_wallet_get_xaddress(self):
-        expected = classic_address_to_xaddress(WALLET.classic_address, None, False)
-        self.assertEqual(WALLET.get_xaddress(), expected)