+++ tests/integration/transactions/test_account_set.py
@@ -1,5 +1,8 @@
 from tests.integration.integration_test_case import IntegrationTestCase
-from tests.integration.it_utils import submit_transaction_async, test_async_and_sync
+from tests.integration.it_utils import (
+    sign_and_reliable_submission_async,
+    test_async_and_sync,
+)
 from tests.integration.reusable_values import WALLET
 from xrpl.models.response import ResponseStatus
 from xrpl.models.transactions import AccountSet
@@ -20,19 +23,16 @@ class TestAccountSet(IntegrationTestCase):
     async def test_required_fields_and_set_flag(self, client):
         account_set = AccountSet(
             account=ACCOUNT,
-            sequence=WALLET.sequence,
             set_flag=SET_FLAG,
         )
-        response = await submit_transaction_async(account_set, WALLET)
+        response = await sign_and_reliable_submission_async(account_set, WALLET, client)
         self.assertEqual(response.status, ResponseStatus.SUCCESS)
         self.assertEqual(response.result["engine_result"], "tesSUCCESS")
-        WALLET.sequence += 1
 
     @test_async_and_sync(globals())
     async def test_all_fields_minus_set_flag(self, client):
         account_set = AccountSet(
             account=ACCOUNT,
-            sequence=WALLET.sequence,
             clear_flag=CLEAR_FLAG,
             domain=DOMAIN,
             email_hash=EMAIL_HASH,
@@ -40,7 +40,6 @@ class TestAccountSet(IntegrationTestCase):
             transfer_rate=TRANSFER_RATE,
             tick_size=TICK_SIZE,
         )
-        response = await submit_transaction_async(account_set, WALLET)
+        response = await sign_and_reliable_submission_async(account_set, WALLET, client)
         self.assertEqual(response.status, ResponseStatus.SUCCESS)
         self.assertEqual(response.result["engine_result"], "tesSUCCESS")
-        WALLET.sequence += 1