+++ tests/integration/transactions/test_deposit_preauth.py
@@ -1,5 +1,8 @@
 from tests.integration.integration_test_case import IntegrationTestCase
-from tests.integration.it_utils import submit_transaction_async, test_async_and_sync
+from tests.integration.it_utils import (
+    sign_and_reliable_submission_async,
+    test_async_and_sync,
+)
 from tests.integration.reusable_values import WALLET
 from xrpl.models.response import ResponseStatus
 from xrpl.models.transactions import DepositPreauth
@@ -13,20 +16,20 @@ class TestDepositPreauth(IntegrationTestCase):
     async def test_authorize(self, client):
         deposit_preauth = DepositPreauth(
             account=ACCOUNT,
-            sequence=WALLET.sequence,
             authorize=ADDRESS,
         )
-        response = await submit_transaction_async(deposit_preauth, WALLET)
+        response = await sign_and_reliable_submission_async(
+            deposit_preauth, WALLET, client
+        )
         self.assertEqual(response.status, ResponseStatus.SUCCESS)
-        WALLET.sequence += 1
 
     @test_async_and_sync(globals())
     async def test_unauthorize(self, client):
         deposit_preauth = DepositPreauth(
             account=ACCOUNT,
-            sequence=WALLET.sequence,
             unauthorize=ADDRESS,
         )
-        response = await submit_transaction_async(deposit_preauth, WALLET)
+        response = await sign_and_reliable_submission_async(
+            deposit_preauth, WALLET, client
+        )
         self.assertEqual(response.status, ResponseStatus.SUCCESS)
-        WALLET.sequence += 1