+++ tests/integration/transactions/test_escrow_cancel.py
@@ -1,5 +1,8 @@
 from tests.integration.integration_test_case import IntegrationTestCase
-from tests.integration.it_utils import submit_transaction_async, test_async_and_sync
+from tests.integration.it_utils import (
+    sign_and_reliable_submission_async,
+    test_async_and_sync,
+)
 from tests.integration.reusable_values import WALLET
 from xrpl.models.response import ResponseStatus
 from xrpl.models.transactions import EscrowCancel
@@ -14,10 +17,11 @@ class TestEscrowCancel(IntegrationTestCase):
     async def test_all_fields(self, client):
         escrow_cancel = EscrowCancel(
             account=ACCOUNT,
-            sequence=WALLET.sequence,
             owner=OWNER,
             offer_sequence=OFFER_SEQUENCE,
         )
-        response = await submit_transaction_async(escrow_cancel, WALLET)
+        response = await sign_and_reliable_submission_async(
+            escrow_cancel, WALLET, client
+        )
         # Actual engine_result is `tecNO_TARGET since OWNER account doesn't exist
         self.assertEqual(response.status, ResponseStatus.SUCCESS)