+++ tests/integration/transactions/test_escrow_finish.py
@@ -1,5 +1,8 @@
 from tests.integration.integration_test_case import IntegrationTestCase
-from tests.integration.it_utils import submit_transaction_async, test_async_and_sync
+from tests.integration.it_utils import (
+    sign_and_reliable_submission_async,
+    test_async_and_sync,
+)
 from tests.integration.reusable_values import WALLET
 from xrpl.models.response import ResponseStatus
 from xrpl.models.transactions import EscrowFinish
@@ -22,13 +25,14 @@ class TestEscrowFinish(IntegrationTestCase):
     async def test_all_fields(self, client):
         escrow_finish = EscrowFinish(
             account=ACCOUNT,
-            sequence=WALLET.sequence,
             owner=OWNER,
             offer_sequence=OFFER_SEQUENCE,
             condition=CONDITION,
             fulfillment=FULFILLMENT,
         )
-        response = await submit_transaction_async(escrow_finish, WALLET)
+        response = await sign_and_reliable_submission_async(
+            escrow_finish, WALLET, client
+        )
         # Actual engine_result will be 'tecNO_TARGET' since using non-extant
         # account for OWNER
         self.assertEqual(response.status, ResponseStatus.SUCCESS)